Move some tests to core
